+++ sctoolbox/tools/frip.py
@@ -10,7 +10,7 @@ from sctoolbox._settings import settings
 
 from beartype import beartype
 
-from typing import Tuple
+from beartype.typing import Tuple
 logger = settings.logger